+++ src/cop701/bot/Wrapper.java
@@ -6,11 +6,11 @@ import java.util.List;
 import cop701.common.Move;
 
 public class Wrapper {
-	int val;
+	double val;
 	List<Move> moves;
 	
 	public Wrapper() {
-		val = 0;
+		val = 0.0;
 		moves = new ArrayList<Move>();
 	}
 }
